+++ R/bcrd3f2.R
@@ -105,8 +105,8 @@ power.bcrd3f2 <- function(es = .25, alpha = .05, two.tailed = TRUE,
 
 cosa.bcrd3f2 <- function(cn1 = 0, cn2 = 0, cn3 = 0, cost = NULL,
                         n1 = NULL, n2 = NULL, n3 = NULL, p = NULL,
-                        n0 = c(10, 100 + g2, 5), p0 = .50,
-                        constrain = "power", round = TRUE,
+                        n0 = c(10, 100 + g2, 5), p0 = .499,
+                        constrain = "power", round = TRUE, max.power = FALSE,
                         local.solver = c("LBFGS", "SLSQP", "MMA", "COBYLA"),
                         rhots = NULL, k1 = -6, k2 = 6, dists = "normal",
                         power = .80, es = .25, alpha = .05, two.tailed = TRUE,
@@ -131,6 +131,10 @@ cosa.bcrd3f2 <- function(cn1 = 0, cn2 = 0, cn3 = 0, cost = NULL,
     stop("Non-logical value for 'two.tailed'", call. = FALSE)
   }
 
+  if(!is.logical(max.power) || length(max.power) > 1){
+    stop("Non-logical value for 'max.power'", call. = FALSE)
+  }
+
   if(any(n2 - g2 < 3)){
     stop("Insufficient sample size, increase 'n2'", call. = FALSE)
   }
@@ -142,21 +146,54 @@ cosa.bcrd3f2 <- function(cn1 = 0, cn2 = 0, cn3 = 0, cost = NULL,
   fun <- "cosa.bcrd3f2"
   lb <- c(1, g2 + 3, 1)
 
-  .df <- quote(n2 - g2 - 2 - (n2 - 2) * (1 - n3))
+  .df <- quote(n3 * (n2 - 2) - g2)
   .sse <- quote(sqrt(d * (rho2 * (1 - r22) / (p * (1 - p) * n3 * n2) +
                             (1 - rho2) * (1 - r21) / (p * (1 - p) * n3 * n2 * n1))))
   .cost <- quote(n3 * cn3 +
                    n3 * n2 * (cn2[2] + p * (cn2[1] - cn2[2])) +
                    n3 * n2 * n1 * (cn1[2] + p * (cn1[1] - cn1[2])))
 
+  .var.jacob <- expression(
+    c(
+      -d * (1 - rho2) * (1 - r21) / (p * (1 - p) * n2 * n3 * n1^2),
+
+      -d * rho2 * (1 - r22) / (p * (1 - p) * n2^2 * n3) -
+        d * (1 - rho2) * (1 - r21) / (p * (1 - p) * n2^2 * n3 * n1),
+
+      -d * rho2 * (1 - r22) / (p * (1 - p) * n2 * n3^2) -
+        d * (1 - rho2) * (1 - r21) / (p * (1 - p) * n2 * n3^2 * n1),
+
+      -(1 - 2 * p) * d * rho2 * (1 - r22) / ((1 - p)^2 * p^2 * n2 * n3) -
+        (1 - 2 * p) * d * (1 - rho2) * (1 - r21) / ((1 - p)^2 * p^2 * n2 * n3 * n1)
+    )
+  )
+
+  .cost.jacob <- expression(
+    c(
+      n3 * n2 * (p * cn1[1] + (1 - p) * cn1[2]),
+
+      n3 * (p * cn2[1] + (1 - p) * cn2[2]) +
+        n3 * n1 * (p * cn1[1] + (1 - p) * cn1[2]),
+
+      cn3 +
+        n2 * (p * cn2[1] + (1 - p) * cn2[2]) +
+        n2 * n1 * (p * cn1[1] + (1 - p) * cn1[2]),
+
+      n3 * n2 * (cn2[1] - cn2[2]) +
+        n3 * n2 * n1 * (cn1[1] - cn1[2])
+    )
+  )
+
   cosa <- .cosa(cn1 = cn1, cn2 = cn2, cn3 = cn3, cost = cost,
-                constrain = constrain, round = round, local.solver = local.solver,
+                constrain = constrain, round = round,
+                max.power = max.power, local.solver = local.solver,
                 power = power, es = es, alpha = alpha, two.tailed = two.tailed,
                 rhots = rhots, k1 = k1, k2 = k2, dists = dists,
                 rho2 = rho2, r21 = r21, r22 = r22,
                 g2 = g2, p0 = p0, p = p, n0 = n0, n1 = n1, n2 = n2, n3 = n3)
   cosa.out <- list(parms = list(cn1 = cn1, cn2 = cn2, cn3 = cn3, cost = cost,
-                                constrain = constrain, round = round, local.solver = local.solver,
+                                constrain = constrain, round = round,
+                                max.power = max.power, local.solver = local.solver,
                                 power = power, es = es, alpha = alpha, two.tailed = two.tailed,
                                 rhots = rhots, k1 = k1, k2 = k2, dists = dists,
                                 rho2 = rho2, r21 = r21, r22 = r22,