+++ R/crd4r4.R
@@ -109,8 +109,9 @@ power.crd4r4 <- function(es = .25, alpha = .05, two.tailed = TRUE,
 # power.crd4r4(es = .683, rho4 = .05, rho3 = .05, rho2 = .10, n1 = 10, n2 = 2, n3 = 3, n4 = 20)
 
 cosa.crd4r4 <- function(cn1 = 0, cn2 = 0, cn3 = 0, cn4 = 0, cost = NULL,
-                        n1 = NULL, n2 = NULL, n3 = NULL, n4 = NULL, p = NULL, n0 = c(10, 3, 100, 5 + g4), p0 = .50,
-                        constrain = "power", round = TRUE,
+                        n1 = NULL, n2 = NULL, n3 = NULL, n4 = NULL, p = NULL,
+                        n0 = c(10, 3, 100, 5 + g4), p0 = .499,
+                        constrain = "power", round = TRUE, max.power = FALSE,
                         local.solver = c("LBFGS", "SLSQP", "MMA", "COBYLA"),
                         rhots = NULL, k1 = -6, k2 = 6, dists = "normal",
                         power = .80, es = .25, alpha = .05, two.tailed = TRUE,
@@ -135,6 +136,10 @@ cosa.crd4r4 <- function(cn1 = 0, cn2 = 0, cn3 = 0, cn4 = 0, cost = NULL,
     stop("Non-logical value for 'two.tailed'", call. = FALSE)
   }
 
+  if(!is.logical(max.power) || length(max.power) > 1){
+    stop("Non-logical value for 'max.power'", call. = FALSE)
+  }
+
   if(any(n4 - g4 < 3)){
     stop("Insufficient sample size, increase 'n4'", call. = FALSE)
   }
@@ -156,8 +161,55 @@ cosa.crd4r4 <- function(cn1 = 0, cn2 = 0, cn3 = 0, cn4 = 0, cost = NULL,
                    n4 * n3 * n2 * (cn2[2] + p * (cn2[1] - cn2[2])) +
                    n4 * n3 * n2 * n1 * (cn1[2] + p * (cn1[1] - cn1[2])))
 
+  .var.jacob <- expression(
+    c(
+      -d * (1 - rho4 - rho3 - rho2) * (1 - r21) / (p * (1 - p) * n2 * n3 * n4 * n1^2),
+
+      -d * rho2 * (1 - r22) / (p * (1 - p) * n2^2 * n3 * n4) -
+        d * (1 - rho4 - rho3 - rho2) * (1 - r21) / (p * (1 - p) * n2^2 * n3 * n4 * n1),
+
+      -d * rho3 * (1 - r23) / (p * (1 - p) * n3^2 * n4) -
+        d * rho2 * (1 - r22) / (p * (1 - p) * n2 * n3^2 * n4) -
+        d * (1 - rho4 - rho3 - rho2) * (1 - r21) / (p * (1 - p) * n2 * n3^2 * n4 * n1),
+
+      -d * rho4 * (1 - r24) / (p * (1 - p) * n4^2) -
+        d * rho3 * (1 - r23) / (p * (1 - p) * n3 * n4^2) -
+        d * rho2 * (1 - r22) / (p * (1 - p) * n2 * n3 * n4^2) -
+        d * (1 - rho4 - rho3 - rho2) * (1 - r21) / (p * (1 - p) * n2 * n3 * n4^2 * n1),
+
+      -(1 - 2 * p) * d * rho4 * (1 - r24) / ((1 - p)^2 * p^2 * n4) -
+        (1 - 2 * p) * d * rho3 * (1 - r23) / ((1 - p)^2 * p^2 * n3 * n4) -
+        (1 - 2 * p) * d * rho2 * (1 - r22) / ((1 - p)^2 * p^2 * n2 * n3 * n4) -
+        (1 - 2 * p) * d * (1 - rho4 - rho3 - rho2) * (1 - r21) / ((1 - p)^2 * p^2 * n2 * n3 * n4 * n1)
+    )
+  )
+
+  .cost.jacob <- expression(
+    c(
+      n4 * n3 * n2 * (p * cn1[1] + (1 - p) * cn1[2]),
+
+      n4 * n3 * (p * cn2[1] + (1 - p) * cn2[2]) +
+        n4 * n3 * n1 * (p * cn1[1] + (1 - p) * cn1[2]),
+
+      n4 * (p * cn3[1] + (1 - p) * cn3[2]) +
+        n4 * n2 * (p * cn2[1] + (1 - p) * cn2[2]) +
+        n4 * n2 * n1 * (p * cn1[1] + (1 - p) * cn1[2]),
+
+      (p * cn4[1] + (1 - p) * cn4[2]) +
+        n3 * (p * cn3[1] + (1 - p) * cn3[2]) +
+        n3 * n2 * (p * cn2[1] + (1 - p) * cn2[2]) +
+        n3 * n2 * n1 * (p * cn1[1] + (1 - p) * cn1[2]),
+
+      n4 * (cn4[1] - cn4[2]) +
+        n4 * n3 * (cn3[1] - cn3[2]) +
+        n4 * n3 * n2 * (cn2[1] - cn2[2]) +
+        n4 * n3 * n2 * n1 * (cn1[1] - cn1[2])
+    )
+  )
+
   cosa <- .cosa(cn1 = cn1, cn2 = cn2, cn3 = cn3, cn4 = cn4, cost = cost,
-                constrain = constrain, round = round, local.solver = local.solver,
+                constrain = constrain, round = round,
+                max.power = max.power, local.solver = local.solver,
                 power = power, es = es, alpha = alpha, two.tailed = two.tailed,
                 rhots = rhots, k1 = k1, k2 = k2, dists = dists,
                 rho2 = rho2, rho3 = rho3, rho4 = rho4,
@@ -165,7 +217,8 @@ cosa.crd4r4 <- function(cn1 = 0, cn2 = 0, cn3 = 0, cn4 = 0, cost = NULL,
                 g4 = g4, p0 = p0, p = p, n0 = n0,
                 n1 = n1, n2 = n2, n3 = n3, n4 = n4)
   cosa.out <- list(parms = list(cn1 = cn1, cn2 = cn2, cn3 = cn3, cn4 = cn4, cost = cost,
-                                constrain = constrain, round = round, local.solver = local.solver,
+                                constrain = constrain, round = round,
+                                max.power = max.power, local.solver = local.solver,
                                 power = power, es = es, alpha = alpha, two.tailed = two.tailed,
                                 rhots = rhots, k1 = k1, k2 = k2, dists = dists,
                                 rho2 = rho2, rho3 = rho3, rho4 = rho4,