+++ NEWS.md
@@ -1,9 +1,14 @@
+### Changes in cosa 1.2.0
+ - Jacobians added 
+ - Option for maximizing power when primary constraint is on total cost
+ - Bug fixes for `cosa.bird2r1()` function 
+
 ### Changes in cosa 1.1.0
  - Experimental wrapper functions `cosa()`, `mdes()` and `power()` added
  - Designs with top-level stratification (fixed blocks) added
  - Bug fixes for `bcrd4r2` and `bcrd4r3` designs
  - Handling of single parameter optimization improved
- - S3 `summary` and `plot` methods added
+ - S3 `summary()` and `plot()` methods added
  - Validity checks for arguments improved
  - Citation added